+++ src/arrayvec.rs
@@ -881,6 +881,17 @@ pub struct IntoIter<T, const CAP: usize, LenType: LenUint = DefaultLenType<CAP>>
     index: usize,
     v: ArrayVec<T, CAP, LenType>,
 }
+impl<T, const CAP: usize> IntoIter<T, CAP> {
+    /// Returns the remaining items of this iterator as a slice.
+    pub fn as_slice(&self) -> &[T] {
+        &self.v[self.index..]
+    }
+
+    /// Returns the remaining items of this iterator as a mutable slice.
+    pub fn as_mut_slice(&mut self) -> &mut [T] {
+        &mut self.v[self.index..]
+    }
+}
 
 impl<T, const CAP: usize, LenType: LenUint> Iterator for IntoIter<T, CAP, LenType> {
     type Item = T;
@@ -1090,7 +1101,9 @@ impl<T, const CAP: usize, LenType: LenUint> ArrayVec<T, CAP, LenType> {
             if let Some(elt) = iter.next() {
                 if ptr == end_ptr && CHECK { extend_panic(); }
                 debug_assert_ne!(ptr, end_ptr);
-                ptr.write(elt);
+                if mem::size_of::<T>() != 0 {
+                    ptr.write(elt);
+                }
                 ptr = raw_ptr_add(ptr, 1);
                 guard.data += 1;
             } else {
@@ -1117,7 +1130,7 @@ impl<T, const CAP: usize, LenType: LenUint> ArrayVec<T, CAP, LenType> {
 unsafe fn raw_ptr_add<T>(ptr: *mut T, offset: usize) -> *mut T {
     if mem::size_of::<T>() == 0 {
         // Special case for ZST
-        ptr.cast::<u8>().wrapping_add(offset).cast()
+        ptr.cast::<u8>().wrapping_add(offset).cast::<T>()
     } else {
         ptr.add(offset)
     }
@@ -1298,5 +1311,39 @@ impl<'de, T: Deserialize<'de>, const CAP: usize, LenType: LenUint> Deserialize<'
         }
 
         deserializer.deserialize_seq(ArrayVecVisitor::<T, CAP, LenType>(PhantomData))
+    }
+}
+
+#[cfg(feature = "borsh")]
+/// Requires crate feature `"borsh"`
+impl<T, const CAP: usize> borsh::BorshSerialize for ArrayVec<T, CAP>
+where
+    T: borsh::BorshSerialize,
+{
+    fn serialize<W: borsh::io::Write>(&self, writer: &mut W) -> borsh::io::Result<()> {
+        <[T] as borsh::BorshSerialize>::serialize(self.as_slice(), writer)
+    }
+}
+
+#[cfg(feature = "borsh")]
+/// Requires crate feature `"borsh"`
+impl<T, const CAP: usize> borsh::BorshDeserialize for ArrayVec<T, CAP>
+where
+    T: borsh::BorshDeserialize,
+{
+    fn deserialize_reader<R: borsh::io::Read>(reader: &mut R) -> borsh::io::Result<Self> {
+        let mut values = Self::new();
+        let len = <u32 as borsh::BorshDeserialize>::deserialize_reader(reader)?;
+        for _ in 0..len {
+            let elem = <T as borsh::BorshDeserialize>::deserialize_reader(reader)?;
+            if let Err(_) = values.try_push(elem) {
+                return Err(borsh::io::Error::new(
+                    borsh::io::ErrorKind::InvalidData,
+                    format!("Expected an array with no more than {} items", CAP),
+                ));
+            }
+        }
+
+        Ok(values)
     }
 }